Move flags to internal/flags
